Add camera and move to /include/
