+++ 5333/build.gradle
@@ -40,8 +40,8 @@ model {
                 lib library: "wpilib"
                 lib library: "ctre"
                 lib library: "navx"
-                lib library: "pathfinder"
                 lib library: "openrio.powerup.matchData"
+                lib library: "pathfinder"
                 lib project: ":common", library: "curtinFrcCommon", linkage: "static"
             }
         }
